# packages/safety-api-zod/src/generated/types/updateTrip.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafeTrip AI API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTripStatus } from './updateTripStatus';

export interface UpdateTrip {
  /** @minLength 1 */
  title?: string;
  destination?: string;
  status?: UpdateTripStatus;
}